chore: stop waiting for aliases in the API
This change introduces a small paradigm change in how aliases are
handled. The AliasAssigned field is not a pointer that is only set once
we know that the alias controller has processed the object. A caller can
check for nil versus false to determine whether the controller has
process the object and the alias will not be assigned.

Signed-off-by: Donnie Adams <donnie@acorn.io>