fix: improve performance for checking for configured model providers
This change simplifies the process for checking for configured model
providers. When listing model providers, instead of making a call to
reveal each credential, this change will make one call to list
credentials.

A side effect is that empty environment variables are not treated as
"configured" while there were not previously. Care is taken to remove
empty environment variables from the credential before saving it.

Signed-off-by: Donnie Adams <donnie@acorn.io>