fix: set the correct credential tool for tool reference
There were a couple issues. First, since we are using the url package,
the credential tool was path-escaped. This change fixes that.

Secondly, the credential set here was not usable by GPTScript if the
reference was of the form "sub tool from path/to/tool". This change also
addresses that by parsing the tool reference.

Signed-off-by: Donnie Adams <donnie@acorn.io>