feat: add support for credential auth for non-tool-references
Support authentication for agent and workflow tools that use
non-tool-references in their definition. That is, a GitHub hosted tool
or a local tool.

During the implementation and testing for non-tool-references, I also
spotted a gap: nested tools. That is, if an agent uses a tool that uses
a tool that has a credential, then the authentication for that tool
would not be processed. After this change, it will.

Signed-off-by: Donnie Adams <donnie@acorn.io>