docs: fetch updated tool authoring readme snippets
Fetch the latest tool authoring README.md snippets from the respective
example tool repos to include the doc updates initially authored by
https://github.com/oskapt in https://github.com/obot-platform/obot/pull/1217.

This is change depends on the following PRs, which must merge before
this:
- https://github.com/obot-platform/go-hash-tool/pull/2
- https://github.com/obot-platform/python-hash-tool/pull/3
- https://github.com/obot-platform/node-hash-tool/pull/2

Note: The docs snippet plugin is too complex and [is slated to be
removed](https://github.com/obot-platform/obot/issues/1236) in the
medium to long-term. This should be the last time we use this flow
for updating the docs.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>
Co-authored-by: Adrian Goins <171292+oskapt@users.noreply.github.com>