enhance: allow tools to indicate which credentials don't need auth
Certain tools have credentials that don't require the use to
authenticate them. For example, the sys.model.provider.credential is
handled by the system and never needs user interaction.

This change introduces the processing of the noUserAuth metadata field
to indicate which credentials don't need user auth. If a credential name
is included in the noUserAuth metadata for a tool, then that credential
name is disregarded for user auth. The sys.model.provider.credential is
included by default for every tool.

Signed-off-by: Donnie Adams <donnie@acorn.io>