feat: add support for OAuth PATs
Instead of configuring and using OAuth, an app can specify that it
supports using personal access tokens. If this is the case, then Obot
will pass an extra environment variable to the oauth credential tool to
indicate which integrations support tokens. If the oauth2 credential
tool should prompt the user for a token instead of using OAuth, then
Obot will not pass the environment variables that feed the URLs to the
tool.

A side effect of this change is that OAuth apps no longer default to
global.

Signed-off-by: Donnie Adams <donnie@acorn.io>