fix: allow model provider definition to be updated
If the model provider tool reference definition is wrong, then it will
never be fixed because returning an error means that the status will not
be properly saved. Reversing the order of the controllers ensures that
the definition can be updated before we try to deal with the model
provider.

Signed-off-by: Donnie Adams <donnie@acorn.io>