fix: ensure threads current and latest runs are active
A previous change made "old" runs inactive. Inactive runs are not cached nor
watched by the controllers.

This wasn't a problem until the cached GET behavior was changed to no longer do
an uncached get when a not-found error occurs. Now we need to ensure that every
thread's latest and current runs are active.

Signed-off-by: Donnie Adams <donnie@acorn.io>