enhance: refactor authn method selection prompts
Change how authentication method selection prompts are displayed in the
user UI so that:
- The submit button isn't displayed
- The field name and "not visible to AI" disclaimer are
  hidden
- The text of the cancel button is changed to "Nevermind, don't
  authenticate"
- Buttons are all the same width and are centered below the description

Note: The logic to use this formatting for the prompt form is predicated
on having a single field named "Authentication Method" that has defined
`options`. This is probably not the best way to discriminate here.
Another option is to add metadata to the prompt and plumb that through.
Would like to see what folks think of this approach first before I go
down that route.

Part of https://github.com/obot-platform/obot/issues/2369

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>