fix: stop using the user ID in the data context for encryption
When a user is created, the ID is 0 until it is inserted into the
database. Therefore, this field is not safe to use in the data context
because it will be different when we try to decrypt it.

This change also includes a force query parameter for the
encrypt-all-users API that will force re-encryption of all users
regardless if those users are already encrypted.

Signed-off-by: Donnie Adams <donnie@acorn.io>