fix: public chatbot user ui
Ensure the editor field is set on project shares for owners of a chatbot
so they can access the chatbot. This also ensures owners never encounter
the warning page and fixes a problem where the warning page doesn't
close after clicking "accept" (note this applied to both users and
owners).

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>