feat: backend support for agent templates
Make backend changes required to support Agent Templates.

This includes a new public-facing `ProjectTemplate` resource and APIs to
interact with it. Under the hood `ProjectTemplates` are implemented as
an extension of the `Thread` and `ThreadShare` storage types, so makes use
of much of the old "copy Obot" logic still in place.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>