enhance: normalize chatbot ux across personas
Normalize ChatBot UX so that agent owners have the same experience as
consumers of their ChatBot. This means that when they access a share
link, they will now interact with a shared "copy" instead of directly
with the parent project.

This commit also includes changes to ensure the ChatBot user warning is
properly closed after the user accepts.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>