fix: add some robustness to deployment-based MCP servers
There are two issues addressed here. First, if the MCP server pod is
restarted without the Obot server being restarted, then existing MCP
server sessions are broken because the MCP server won't recognize them.
This change addresses that by using the pod name in as the scope for the
MCP server client, ensuring that a new session is created if the pod
restarts. New permissions are added to the service account role for
this.

Secondly, a bump to gptscript fixes an issue where the Digest of a
server config could change when Obot restarts, even if the server
configuration hadn't changed. This would leave orphaned deployments with
no way of cleaning them up.

Signed-off-by: Donnie Adams <donnie@acorn.io>