fix: allow MCP servers with no tools allowed
Previously, this change would not distinguish between allowedTools being
nil and having length zero, interpretting both to mean that all tools for
the MCP server to be allowed.

Now, if the allowedTools is non-nil and length zero, then this will be
treated as not allowing any tools from the MCP server.

Signed-off-by: Donnie Adams <donnie@acorn.io>