enhance: ensure that MCP servers used with chat go through the gateway
In this change, the UI will construct an MCP server that goes through
the gateway for chatting. That way, regardless of what kind of MCP
server is being used, all the config is passed through.

To achieve this, a special token is provided to such MCP servers so
that authentication and authorization are achieved properly.

Signed-off-by: Donnie Adams <donnie@acorn.io>