fix: disable copy button when clipboard is unavailable
`navigator.clipboard` is unavailable in insecure contexts (like
non-localhost http). This results in the copy button not working.

Detect this scenario and disable the copy button.

Addresses https://github.com/obot-platform/obot/issues/3433

Note: I'm not sure this is actually how we want to handle this. I can
try another method to make this work in these contexts, but wanted to
gather feedback on this approach first.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>