fix: wait for all capabilities to load before displaying them
Wait for requests to get tools, prompts, and resources to finish before
displaying MCP server capabilities to a user.

This ensure that even if a 424 (Failed Dependency) is returned by one of
the requests, it doesn't prevent the UI from showing the results of the
others to user in the capabilities section.

Addresses https://github.com/obot-platform/obot/issues/3492

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>