fix: support sqlite3 for audit logs
There are two issues fixed here. The first is that Go uses local
timezone by default, and the sqlite package we use respects this.
Therefore, when listing audit logs and mcp stats, use the local time so
sqlite does the right thing.

Secondly, ILIKE is only a postgres feature, so this change will use LIKE
for other dialects.

Signed-off-by: Donnie Adams <donnie@acorn.io>