fix: alleviate block when deleting users when using sqlite
When we use sqlite, we have a single database connection. Also, sqlite's
locking is not a sophisticated as others. So, when we're deleting a
user in a transaction, that locks the database. Then, we're blocked
when we try to clean up other resources associated with the user from
inside the transaction.

This change alleviates that blocking by cleaning up all user-based
resources in the user-delete controller.

Signed-off-by: Donnie Adams <donnie@acorn.io>