fix: stop closing OAuth clients
A previous implementation closed the client we used to check for OAuth
because we didn't want this extra client lying around. It would cause
problems down the road.

A recent change separated clients by session. The upshot is that we no
longer need to close this OAuth client and it can remain active to make
future checks quicker.

There was a bug here as well: the ShutdownServer call was shutting down
all clients for a given server instead of just the OAuth client.

Signed-off-by: Donnie Adams <donnie@acorn.io>