enhance: validate auth provider options on configure
Call the target auth provider's `validate` tool on the given set of
options before (re)configuring the provider.

If the call to the validation tool fails or returns a validation error,
abort (re)configuration and bubble up the error details with a `422 Unprocessible Entity`
response.

If the auth provider does not have an associated `validate tool, treat
the given options as valid.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>