fix: remove edit mode component
The `EditMode` component periodically updates the focused project if it's changed.
Upgrading a project from a share will also result in the project being
updated out-of-band in the backend. This can sometimes cause the update
issued by `EditMode` to fail with a conflict and result in an error
message being presented to the user.

Since we've already switched over to a manual `Update` button for
changing project configuration, we can drop this component and prevent
the failure from occurring.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>