feat: add new auditor an owner roles (#4319)
This change introduces two new roles: owner and auditor. Owners are
the same as admins except that owners can grant/revoke the owner
and auditor roles. A user must be an auditor in order to see the request
and response information for audit logs, as well as see the chat history
for threads. Admins (and owners) now have more limited scopes of what
they can see.

This change includes a rewrite of the role and authorization framework
in Obot. Authorization is more role-based and users can have more than
one role (like admin and auditor). Roles are migrated once on startup.

Lastly for the backend, there is a new configuration parameter to set
explicit owners, similar to the same for admins.

The UI also incorporates these new roles and allows owners to grant or
revoke the owner and auditor roles. Additionally, the handoff from the
bootstrap user is done after a user is granted an owner role rather than an
admin role.

Signed-off-by: Donnie Adams <donnie@obot.ai>
Co-authored-by: Ivy <ivy.jeong@gmail.com>