fix: remove / from MCP server names for nanobot
Nanobot expects the server name to not have any forward-slashes in it because it
uses forward-slashes for tool mapping. So, if the server name is something like
@modelcontextprotocol/server-sequential-thinking, then nanobot will parse this
as a server with name @modelcontextprotocol and a tool with name
server-sequential-thinking.

This change will replace / with - and nanobot will correctly parse the config.

Signed-off-by: Donnie Adams <donnie@obot.ai>