fix: treat owners and admins separately when removing (#4819)
We deleting or demoting a user that is an ownwer, we would allow it if there was
another admin in the system. This is incorrect because it would be impossible
for any users to be promoted to owner.

This change separates the owner check from the admin check to ensure there is
always at least one owner.

Signed-off-by: Donnie Adams <donnie@obot.ai>