fix: always call configure endpoint when launching composites
When a component of a composite MCP server that requires config is
disabled, but there are no other components that require configuration,
the UI skips calling the configure endpoint for the composite. However,
the configure endpoint is also used to disable MCP servers, so when this
step is skipped, the server is never actually disabled and the
subsequent call to the launch endpoint fails.

To fix this, always call the configure endpoint when launching composite
MCP servers.

Addresses https://github.com/obot-platform/obot/issues/4872

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>