fix: correctly determine MCP server ID for users that have yet to login
If a user was logged out when they attempted to use a connect URL, there were
two things that would go wrong:
1. An MCP server for the "anonymous" user was created.
2. The user would not be able to connect to the MCP server (because it belonged
to the "anonymous" user).

This change addresses both of these by checking the for the MCP server ID only
after we know the user has been authenticated (in the callback handler). The
erroneously created servers for the anonymous user are now deleted.

There is a corresponding bug fixed here where the connect ID having 5 parts
would cause 404 errors in nanobot.

Signed-off-by: Donnie Adams <donnie@obot.ai>