fix: stop failing if component server catalog entry is deleted (#5159)
For composite servers, a user is allowed to delete the catalog entry for a
component server and have the composite server still function properly. If the
catalog entry no longer exists, then it is not possible to retrieve the catalog
ID from the entry. This change adds a field to the status to track the catalog
ID of the component server, which was previously pulled from the catalog entry.

Signed-off-by: Donnie Adams <donnie@obot.ai>