enhance: add idle MCP server shutdown
After this change, idle MCP servers will be automatically shutdown on the
configured interval. The activity of each server will be tracked at the
granularity of 1/4 of the configured interval or every 15 minutes, whichever is
smaller. That is, if the shutdown interval is 30 minutes, the activity will be
tracked every 7.5 minutes.

Signed-off-by: Donnie Adams <donnie@obot.ai>